+++ indico/modules/events/registration/client/js/form/fields/SingleChoiceInput.jsx
@@ -8,11 +8,12 @@
 import createDecorator from 'final-form-calculate';
 import _ from 'lodash';
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, {useMemo} from 'react';
 import {Field} from 'react-final-form';
 import {useSelector} from 'react-redux';
 import {Form, Label, Dropdown} from 'semantic-ui-react';
 
+import {ComboBox} from 'indico/react/components';
 import {FinalCheckbox, FinalDropdown, FinalField, parsers as p} from 'indico/react/forms';
 import {Translate} from 'indico/react/i18n';
 
@@ -31,6 +32,8 @@ function SingleChoiceDropdown({
   existingValue,
   value,
   onChange,
+  onFocus,
+  onBlur,
   disabled,
   isRequired,
   isPurged,
@@ -42,23 +45,31 @@ function SingleChoiceDropdown({
   const management = useSelector(getManagement);
   const currency = useSelector(getCurrency);
   const selectedChoice = choices.find(c => c.id in value) || {};
+  const selectedSeats = value[selectedChoice.id] || 0;
 
   const isPaidChoice = choice => choice.price > 0 && paid;
   const isPaidChoiceLocked = choice => !management && isPaidChoice(choice);
 
   let extraSlotsDropdown = null;
-  if (withExtraSlots && selectedChoice && selectedChoice.maxExtraSlots > 0) {
-    const options = _.range(1, selectedChoice.maxExtraSlots + 2).map(i => ({
-      key: i,
+  const extraSlotsLabelId = `${id}-label`;
+  const shouldShowExtraSlots = withExtraSlots && selectedChoice && selectedChoice.maxExtraSlots > 0;
+  const shouldShowExtraSlotsLabel = shouldShowExtraSlots && !!selectedChoice.price;
+
+  if (shouldShowExtraSlots) {
+    const slotValues = _.range(1, selectedChoice.maxExtraSlots + 2);
+    const options = slotValues.map(i => ({
       value: i,
-      text: i,
       disabled:
         selectedChoice.placesLimit > 0 &&
         (placesUsed[selectedChoice.id] || 0) - (existingValue[selectedChoice.id] || 0) + i >
           selectedChoice.placesLimit,
     }));
+    const comboBoxExtraProps = {};
+    if (shouldShowExtraSlotsLabel) {
+      comboBoxExtraProps['aria-describedby'] = extraSlotsLabelId;
+    }
     extraSlotsDropdown = (
-      <Form.Select
+      <ComboBox
         id={id ? `${id}-extraslots` : ''}
         options={options}
         disabled={
@@ -68,73 +79,85 @@ function SingleChoiceDropdown({
             (placesUsed[selectedChoice.id] || 0) - (existingValue[selectedChoice.id] || 0) >=
               selectedChoice.placesLimit)
         }
-        value={value[selectedChoice.id]}
-        onChange={(_evt, data) => onChange({[selectedChoice.id]: data.value})}
-        fluid
+        value={String(selectedSeats)}
+        onChange={evt => {
+          const selectedSlots = Number(evt.target.value);
+          onChange({[selectedChoice.id]: selectedSlots});
+        }}
+        aria-label={Translate.string('Select extra slots')}
+        required
+        {...comboBoxExtraProps}
       />
     );
   }
 
-  const options = choices.map(c => ({
-    key: c.id,
-    value: c.id,
-    disabled:
-      !c.isEnabled ||
-      isPaidChoiceLocked(c) ||
-      (c.placesLimit > 0 && (placesUsed[c.id] || 0) >= c.placesLimit && !existingValue[c.id]),
-    text: c.caption,
-    content: (
-      <div styleName="dropdown-text">
+  const options = choices.map(choice => ({
+    value: choice.caption,
+    label: (
+      <div styleName="dropdown-text" key={choice.id}>
         <div styleName="caption">
-          <ChoiceLabel choice={c} management={management} paid={isPaidChoice(c)} />
+          <ChoiceLabel choice={choice} management={management} paid={isPaidChoice(choice)} />
         </div>
         <div styleName="labels">
-          {!!c.price && (
+          {!!choice.price && (
             <Label>
-              {c.price} {currency}
+              {choice.price} {currency}
             </Label>
           )}
-          {c.placesLimit === 0 ? null : (
+          {choice.placesLimit === 0 ? null : (
             <PlacesLeft
-              placesLimit={c.placesLimit}
-              placesUsed={placesUsed[c.id] || 0}
-              isEnabled={c.isEnabled}
+              placesLimit={choice.placesLimit}
+              placesUsed={placesUsed[choice.id] || 0}
+              isEnabled={choice.isEnabled}
             />
           )}
         </div>
       </div>
     ),
+    disabled:
+      !choice.isEnabled ||
+      isPaidChoiceLocked(choice) ||
+      (choice.placesLimit > 0 &&
+        (placesUsed[choice.id] || 0) >= choice.placesLimit &&
+        !existingValue[choice.id]),
   }));
+
+  const handleChange = evt => {
+    if (!evt.target.value) {
+      onChange({'': -1});
+      return;
+    }
+    const selected = choices.find(c => c.caption === evt.target.value);
+    onChange(selected ? {[selected.id]: 1} : {'': -1});
+  };
 
   return (
     <Form.Group styleName="single-choice-dropdown">
       <Form.Field>
-        <Dropdown
+        <ComboBox
           id={id}
-          selection
-          style={{width: 500, opacity: 1}}
-          placeholder={Translate.string('Choose an option')}
+          onChange={handleChange}
           options={options}
+          value={(!isPurged && selectedChoice.caption) || ''}
+          required={isRequired}
           disabled={disabled}
-          value={(!isPurged && selectedChoice.id) || ''}
-          onChange={(e, data) => onChange(data.value ? {[data.value]: 1} : {})}
-          clearable={!isRequired}
-          search
-          selectOnNavigation={false}
-          selectOnBlur={false}
+          onFocus={onFocus}
+          onBlur={onBlur}
         />
       </Form.Field>
-      {extraSlotsDropdown}
-      {extraSlotsDropdown && !!selectedChoice.price && (
-        <Label pointing="left">
-          {Translate.string('Total: {total} {currency}', {
-            total: (
-              (selectedChoice.extraSlotsPay ? value[selectedChoice.id] : 1) * selectedChoice.price
-            ).toFixed(2),
-            currency,
-          })}
-        </Label>
-      )}
+      <div styleName="extra-slots">
+        {extraSlotsDropdown}
+        {shouldShowExtraSlotsLabel && (
+          <Label pointing="left" id={extraSlotsLabelId}>
+            {Translate.string('Total: {total} {currency}', {
+              total: (
+                (selectedChoice.extraSlotsPay ? selectedSeats : 1) * selectedChoice.price
+              ).toFixed(2),
+              currency,
+            })}
+          </Label>
+        )}
+      </div>
     </Form.Group>
   );
 }
@@ -143,6 +166,8 @@ SingleChoiceDropdown.propTypes = {
   id: PropTypes.string.isRequired,
   value: PropTypes.object.isRequired,
   onChange: PropTypes.func.isRequired,
+  onFocus: PropTypes.func.isRequired,
+  onBlur: PropTypes.func.isRequired,
   disabled: PropTypes.bool.isRequired,
   isRequired: PropTypes.bool.isRequired,
   isPurged: PropTypes.bool.isRequired,
@@ -299,6 +324,8 @@ function SingleChoiceInputComponent({
   existingValue,
   value,
   onChange,
+  onFocus,
+  onBlur,
   disabled,
   isRequired,
   isPurged,
@@ -315,6 +342,8 @@ function SingleChoiceInputComponent({
         value={value}
         existingValue={existingValue}
         onChange={onChange}
+        onFocus={onFocus}
+        onBlur={onBlur}
         disabled={disabled}
         isRequired={isRequired}
         isPurged={isPurged}
@@ -370,6 +399,27 @@ export default function SingleChoiceInput({
   placesUsed,
 }) {
   const existingValue = useSelector(state => getFieldValue(state, fieldId)) || {};
+
+  const optionSet = useMemo(() => new Set(choices.map(choice => choice.id)), [choices]);
+
+  function validate(value) {
+    const noValue = !value || !Object.keys(value).length;
+    if (isRequired && noValue) {
+      return Translate.string('This field is required');
+    }
+
+    if (noValue) {
+      // When there is no value but the field is not required, it's a pass
+      return;
+    }
+
+    for (const key in value) {
+      if (!optionSet.has(key)) {
+        return Translate.string('Please select one of the provided choices');
+      }
+    }
+  }
+
   return (
     <FinalField
       id={htmlId}
@@ -378,11 +428,7 @@ export default function SingleChoiceInput({
       format={v => v || {}}
       required={isRequired}
       isRequired={isRequired}
-      validate={v =>
-        isRequired && (!v || !Object.keys(v).length)
-          ? Translate.string('This field is required.')
-          : undefined
-      }
+      validate={validate}
       disabled={disabled}
       isPurged={isPurged}
       itemType={itemType}